+++ src/endobox/Box.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of endobox.
  *
- * (c) 2015-2019 YouniS Bensalah <younis.bensalah@gmail.com>
+ * (c) 2015-2020 Younis Bensalah <younis.bensalah@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -45,6 +45,9 @@ class Box implements Renderable, \IteratorAggregate
 
     public function __invoke($arg) : Box
     {
+        if (\is_array($arg)) {
+            return $this->assign($arg);
+        }
         return $this->append($arg);
     }