+++ src/endobox/BoxIterator.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of endobox.
  *
- * (c) 2015-2019 YouniS Bensalah <younis.bensalah@gmail.com>
+ * (c) 2015-2020 Younis Bensalah <younis.bensalah@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.