+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased]
+
 ### Added
+- Invoke `Box` with `array` to assign data.
+
 ### Changed